requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cachetools==5.5.0
click==8.1.7
fastapi==0.112.2
fastapi-pagination==0.12.26
GPUtil==1.4.0
greenlet==3.0.3
h11==0.14.0
idna==3.8
nvidia-ml-py==12.535.161
nvitop==1.3.2
psutil==6.0.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
setuptools==74.0.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
termcolor==2.4.0
typing_extensions==4.12.2
uvicorn==0.30.6
